remove console

+++ src/components/redux/AuthenticationSlice.js
@@ -16,7 +16,6 @@ const CURRENT_USER_URL = 'http://127.0.0.1:4000/api/v1/users/current_user';
 
 export const signUp = createAsyncThunk('user/signup', async (newUser) => {
   const response = await axios.post(`${SIGNUP_URL}`, newUser);
-  console.log('Sign up response:', response);
   sessionStorage.setItem('authToken', response.headers.authorization);
   sessionStorage.setItem('isAuthenticated', 'true');
   return response.data;
@@ -25,38 +24,37 @@ export const signUp = createAsyncThunk('user/signup', async (newUser) => {
 export const logIn = createAsyncThunk('user/login', async (newSession) => {
   const response = await axios.post(`${LOGIN_URL}`, newSession);
   sessionStorage.setItem('authToken', response.headers.authorization);
-  console.log('Logged in. Setting isAuthenticated to true');
   sessionStorage.setItem('isAuthenticated', 'true');
   return response.data.status;
 });
+
 export const logOut = createAsyncThunk('user/logout', async () => {
   const authToken = sessionStorage.getItem('authToken');
 
-  const response = await axios.delete(LOGOUT_URL, {
-    headers: {
-      Authorization: authToken,
-    },
-  });
-  console.log('Logout response:', response);
-  sessionStorage.removeItem('authToken');
-  sessionStorage.removeItem('isAuthenticated');
+  if (authToken) {
+    await axios.delete(LOGOUT_URL, {
+      headers: {
+        Authorization: authToken,
+      },
+    });
+
+    sessionStorage.removeItem('authToken');
+    sessionStorage.removeItem('isAuthenticated');
+  }
+
+  return null;
 });
 
 export const fetchCurrentUser = createAsyncThunk(
   'authentication/fetchCurrentUser',
   async () => {
-    try {
-      const authToken = sessionStorage.getItem('authToken');
-      const response = await axios.get(`${CURRENT_USER_URL}`, {
-        headers: {
-          Authorization: authToken,
-        },
-      });
-      return response.data;
-    } catch (error) {
-      console.error('Error fetching current user:', error);
-      throw error;
-    }
+    const authToken = sessionStorage.getItem('authToken');
+    const response = await axios.get(`${CURRENT_USER_URL}`, {
+      headers: {
+        Authorization: authToken,
+      },
+    });
+    return response.data;
   },
 );
 
@@ -86,7 +84,6 @@ const authenticationSlice = createSlice({
         state.isLoading = false;
         state.isAuthenticated = true;
         state.user_id = action.payload.id;
-        console.log('User ID:', action.payload.id);
       });
   },
 });
